INSIGHT - WIP. Insight needs scenarios tests to ensure signal logic working, given the test DXY file.
OEMS - WIP.
OEMS needs scenarios tests to ensure signal logic working, given the test DXY file.

+++ src/main/java/core/service/insight/InsightData.java
@@ -42,7 +42,7 @@ public class InsightData extends SelfDescribingMarshallable {
 
     // oems q real-time data
     public long openOrderId;
-    public double openOrderTimestamp;
+    public long openOrderTimestamp;
     public String openOrderExpiry;
     public String openOrderState;
     public double openOrderSLPrice;
